Alter profile version range in migration class. Display errors when profile loading/saving fails.
Joystick profiles with a version older than 5 can now be migrated.
Joystick profiles with a version as low as version 2 will
now run through the migration class. I have tested it with some
old version 2 profiles I have and it works. I believe version 2
was the first version to save sets so the joystick should
just use the default value for any element that was not
available in an older version.