Fixed issue with old profiles not being usable by the program.
The profile migration class was flawed since old profiles
with a lower configversion than the most recent spec would
cause a migration to be started. However, if the configversion
of a profile was version 5 or older, the old data was discarded
with nothing to replace it so no bindings would be loaded into the
program.