Force resetting XTest pointer acceleration settings.
For now, just change the XTest mouse pointer settings to the desired
settings. Maybe later the original XTest pointer settings should be
saved and reverted when closing the program. When this program first
started, the distro configuration I used at that time did not have
acceleration set for the XTest pointer.