make it possible to build reproducibly
By default, the date and time of the build are embedded into the
binary as sort of origin/version/etc. marker. This is not, however,
reproducible. This patch makes it so that packagers can define a
præprocessor macro, ANTIMICROX_PKG_VERSION, to a C string which is
then used instead; if this macro is not defined nothing changes.

The diff comes from Debian; by upstreaming the new string can be
translated.