Simplify FlashButtonWidget
All child classes of FlashButtonWidget were explicitely listed in
the mainwindow's stylesheet. Hence, new child classes are not correctly
styled until they are added to the stylesheet as well.

As simplification, only add FlashButtonWidget itself to the main stylesheet
so that all child classes are automatically styled correctly.
Editing the stylesheet with QtDesigned caused the file to be
reformatted. Furthermore, remove the unused "flashed" signals.