Fix QTime::restart deprecation warning
Replace it with QElapsedTimer as the deprecation warning suggests.
Remove unused function JoyButton::copyExtraAccelerationState which was
unused since it was introduced in commit f9058369.

In the JoyButtonSlot copy constructor remove the hack that copies the
time via an intermediate string and replace it with a new timer because
setting the time of a QElapsedTimer is not supported.
The new timer is started when the source timer is running.
This should not be a problems because the easingTime is usually very
small and there are not used virtual inputs while users change settings.