Fix QHashIterator::hasPrevious() deprecation warnings
QHashIterator::hasPrevious() is deprecated.
Replace most Java style iterators which use hasPrevious() with STL style
iterators and move the first iteration before the loop. According to the
Qt documentation [1], this should also be more efficient.

Keep the usage of hasPrevious() for JoyButton::slotiter since it is
used in a more sophisticated way. This is also a QListIterator where
hasPrevious is not deprecated.

[1]: https://doc.qt.io/qt-5/containers.html