Fix "jumping" JoyControlStickEditDialog layout
The layout of the JoySensorEditDialog jumps around when the number of
decimals in the bearing and safezone labels change.

Fix it by putting all stick labels and values from the left column in a
QGridLayout and set the length of the first value to the length of a
dummy string with nine decimals.