Implement new stick calibration
Implement an interactive calibration workflow which
automatically detects when enough data was collected by evaluation
statistic properties. The current estimated calibration values are shown
live in red and turn black if the relative error is below 1%.
If not enough values are collected after 30 seconds, time out and
continue with the available data. This avoids long calibration times
that may lead to bug reports because users think it hangs in case of
noisy devices. Determine the final calibration values with a linear regression.

Rework the calibration dialog, e.g remove redundant ProgressBars and
reduce empty space. The new dialog is also ready to be extended for
sensor calibration.

Encode the device index with a bitmask into the type int since this
allows using the QVariant::int user data in the combo box directly.