Add sensors to InputDaemon
Enable sensor events in input daemon, handle them in all input
passes and finally activate them at the end of the last pass.

Set "pending" status bits in firstInputPass.
I'm not sure why exactly this is necessary since it worked well without
it but other analog inputs like axes and sticks do it.
So do it for sensors as well.

Add stub functions for all new interfaces required by the InputDaemon changes
as well as a stub JoySensor class.