Show raw sensor values in Properties dialog
Add a new group for sensors in the right column of the dialog and add
show raw sensor values on progress bar indicators similar to axes.
Store the pointers to the progress bars in member variables because
it is not possible to connect them directly to the three dimensional
sensor "moved" events.

Add global constants for sensor minima and maxima. Since sensors
measure floating point physical parameters there are no real minima and maxima.
So use +/-90 m/s^2 as maximum for accelerometer and +/- 360 °/s for
gyroscope as senseful limits.

Implement sensor event activation in InputDevice and add minimal
implementation of the JoySensor side.