Add JoySensorButton class
A specialized JoyButton subclass is necessary to store a pointer to the
JoySensor, to perform specialized mouse calculations and to store the
new 3D direction.
Add two additional subclasses for acceleromters and gyroscopes which
handle the translated direction name of the button.

Encode the 3D directions into a bitfield. This makes it easier to check
for combined three dimensional directions later. Furthermore, add
dummies for the required JoySensor interfaces.