Add sensor preview widget
Draw sensor position like an artifical horizon instrument.
This should be very intuitive to understand for accelerometer and
not much more difficult for gyroscope. Only draw the diagonal zones for
the XY plane to avoid cluttering.

Add the required interfaces to JoySensor. Use dummies for non-trivial
interfaces.

This widget will be used in the coming sensor edit dialog and
calibration dialog.