Implement JoySensorEditDialog
Show this dialog based on JoyControlStickEditDialog when
clicking on the central sensor button in the GUI. It allows editing sensor
parameters with sliders and spin boxes.

This is the first time in the patcheset where sensors can be watch in
action in the GUI. Show raw and processed sensor values with correct units
in indicators.

Presets and sensor preview widget will be implemented in a later commit.