Implement sensor presets
Add presets for the existing mouse variants to gyroscope.
Add presets for existing keyboard variants to accelerometer.
Do not add all presets to both because a accelerometer mouse or a
gyroscope keyboard does not make too much sense.
Define the preset values in the caonstructor and not in the .ui file
to realize dynamic entries based on the sensor type.
With this commit, preset selection from the sensor edit dialog works.

Factor out a UI thread helper similar to
joycontrolstickcontextmenuhelper or joycontrolstickeditdialoghelper
which will be used used for both types of UI.