Add modifier zone variables to JoyControlStick class
Add member variables for the modifier zone range and its direction
(inside or outside). Set the default values for the new variables so
that the behaviour of existing profiles does not change.
Handle the new variables in basic functions like initialization and
assignment.

The new variables will be evaluated in a following commit.