Implement calibration storage in InputDeviceCalibration
Store modified calibration values for sticks and gyroscope from InputDevice
in the new storage backend InputDeviceCalibration. This new class
internally stores the calibration values in a QHash which is indexed by the
unique device ID string. The individual CalibrationData structs for a
controller are stored in a QList. If a CalibrationData
structure for a given controller, type and input ID already exists when
calibration values are updated, replace it. Otherwise, insert it to the
list.