Implement calibration data XML serialization/deserialization
This allows storing calibration values for another controller which is
very useful when you have a game console with multiple controllers.
Every time you grab a random controller to use it with AntimicroX there
is
a good chance that the calibration values are now incorrect since it is
another physical device.

During XML serialization, all known calibration values are written to
XML and get tagged with the unique device ID string.
During XML deserialization, load all calibration values into the
internal storage and the calibration which unique ID mathces the
connected controller gets applied to the InputDevice.

Calibration values which are not tagged with any device ID get applied
to the currently connected device for easy migration.