Deep-copy JoyButtonSlot's mix slots
When button assignments are copied from one set to another, the shallow
copy of the mix slots causes a crash on exit due to the double free.

Use the already existing but commented out deep-copy code and extract a
method which is used for the copy-constructor and the assignment
operator.

Fixes #454